fix(review): preserve manifest io authority
