+++ tests/infra/test_schema_read_convention.py
@@ -16,7 +16,9 @@ import pytest
 pytestmark = [pytest.mark.layer("infra"), pytest.mark.medium]
 
 _SHARED_READ_SIDE_VALIDATORS = {
-    "read_installed_plugin_artifact_identity": "src/autoskillit/core/_plugin_cache.py",
+    "read_installed_plugin_artifact_identity": (
+        "src/autoskillit/core/_plugin_artifact_identity.py"
+    ),
     "read_projected_plugin_identity": ("src/autoskillit/workspace/_projection_cache.py"),
 }
